+++ examples/network_sets.py
@@ -19,13 +19,14 @@ from pprint import pprint
 from hpOneView.oneview_client import OneViewClient
 from config_loader import try_load_from_file
 
+# To run this example fill the ip and the credentials below or use a configuration file
 config = {
-    "ip": "<ov_ip>",
+    "ip": "<oneview_ip>",
     "credentials": {
         "userName": "<username>",
         "password": "<password>"
     },
-    "api_version": 1200
+    "api_version": "<api_version>"
 }
 
 options = {